Добавил схему взаимодействия

+++ src/main/java/ru/bookstore/rest/BookController.java
@@ -3,10 +3,12 @@ package ru.bookstore.rest;
 
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
-import io.swagger.annotations.ApiResponse;
-import io.swagger.annotations.ApiResponses;
+import org.hibernate.ObjectNotFoundException;
+import org.hibernate.exception.ConstraintViolationException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
+import org.springframework.transaction.annotation.Propagation;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
 import ru.bookstore.domain.Book;
 import ru.bookstore.repositories.BookRepository;
@@ -46,7 +48,7 @@ public class BookController {
     }
 
 
-    @ApiOperation(value = "Метод редактирования информации о книги", response = Book.class, tags = "updateBook")
+    @ApiOperation(value = "Метод редактирования информации о книгию Нужно передать JSON Book", response = Book.class, tags = "updateBook")
     @PutMapping("/books/update/{bookId}")
     public ResponseEntity<Book> updateBook(@PathVariable("bookId") long bookId, @RequestBody(required = true) Book book) {
         book.setId(bookId);
@@ -54,10 +56,15 @@ public class BookController {
         return ResponseEntity.ok().body(book);
     }
 
-    @ApiOperation(value = "Метод удаления книги", tags = "deleteBook")
-    @DeleteMapping("/books/delete/{bookId}")
-    public ResponseEntity deleteBook(@PathVariable("bookId") long bookId) {
-        bookRepository.deleteById(bookId);
+    @Transactional(propagation = Propagation.REQUIRED, rollbackFor = {ObjectNotFoundException.class, ConstraintViolationException.class})
+    @ApiOperation(value = "Метод удаления книг(и).Нужно передать JSON массив из id книг, удаляемых из БД", tags = "deleteBook")
+    @DeleteMapping("/books/delete")
+    public ResponseEntity deleteBook( @RequestBody(required = true) List<Long> bookIds) {
+        if (bookIds != null) {
+            for (Long bookId : bookIds) {
+                bookRepository.deleteById(bookId);
+            }
+        }
         return ResponseEntity.ok().build();
     }
